The changes here aim to improve Bodhi's JSON API response times.
At database/sqlalchemy level, I've changed some problematic relationships load modes. Some were set to be `lazy='joined'`, but that caused to load very large amounts of data from the db.
In some cases I've set the relationship to use inner joins (where the referenced object id was not nullable).

The PR include a database change to set user_id in comments table to be not nullable (I've checked the db and all comments are correctly linked to a user).
There could be some other places where this seems possible: for example all comments should be linked to an update... but there are some in the db for which this isn't true (don't know why).

At object serialization level, I've added a `__relation_fields__` property to BodhiBase object along with relevant object. The scope is to select what attributes should be serialized when the object is loaded as relation inside another object.

# Performance tests
I've made some dummy tests to measure performance improvements of this change.
## Query times
The code used to conduct this test is (from bodhi-shell):
```
def query_time(query):
    t0=time.time()
    query.all()
    print(time.time() - t0)

models = (m.Build, m.Comment, m.BuildrootOverride, m.Package, m.Release, m.Update, m.User)
for model in models:
    query = s.query(model)
    query = query.limit(1000)
    query_time(query)
```
The results are:
| Class | Before changes | After changes | Improvement |
|---|---|---|---|
| Build | 1,51109004020691 | 0,182040214538574 | -87,95% |
| Comment | 0,390185356140137 | 0,038461446762085 | -90,14% |
| BuildrootOverride | 1,15033030509949 | 0,199055910110474 | -82,70% |
| Package | 0,037990093231201 | 0,036218881607056 | -4,66% |
| Release | 0,005536556243896 | 0,005467653274536 | -1,24% |
| Update | 1,27048087120056 | 0,099624633789063 | -92,16% |
| User | 8,3333420753479 | 0,02573299407959 | -99,69% |

## JSON lists
| URL | Before changes | After changes | Improvement |
|---|---|---|---|
| http://localhost:6543/builds/?rows_per_page=1000 | 13s | 4,6s | -64,62% |
| http://localhost:6543/comments/?rows_per_page=1000 | 55s | 31s | -43,64% |
| http://localhost:6543/overrides/?rows_per_page=1000 | 6,1s | 0,6s | -90,16% |
| http://localhost:6543/packages/?rows_per_page=1000 | 186ms | 170ms | -8,60% |
| http://localhost:6543/releases/?rows_per_page=1000 | 75ms | 103ms | +37,33% |
| http://localhost:6543/updates/?rows_per_page=1000 | 136s | 26s | -80,88% |
| http://localhost:6543/users/?rows_per_page=1000 | 11s | 2,5s | -77,27% |

Because the JSON schema is changed, some tools that depends on this can break. So, I've tagged the PR as backwards incompatible.

Some longstanding issues that should be fixed by this:
fix #271
fix #301 
fix #1899
fix #3204 


Signed-off-by: Mattia Verga <mattia.verga@tiscali.it>