@cverna 2020-05-18
Overall this is looking good, I just want to check the fedora-messaging message sent, since I think this is using the json serialization. If that's the case we need to check the messages schema and also be noisy about the changes in case someone relies on a field that we are going to remove.