@mattiaverga 2020-05-19
> Overall this is looking good, I just want to check the fedora-messaging message sent, since I think this is using the json serialization. If that's the case we need to check the messages schema and also be noisy about the changes in case someone relies on a field that we are going to remove.

Well, while writing this PR I gave a look to some tools that use Bodhi Json APIs, like fedora-easy-karma and qa-stats and I think they should not be broken by this.
About message schemes, I hope the unit tests we run are reliable enough...