@cverna 2020-05-25
Ok so this impacts the messages sent. I did a bit of checking in vagrant and for example

The bodhi.update.edit message in prod https://apps.fedoraproject.org/datagrepper/id?id=2020-b28bbe36-4715-48b7-94bc-dd70d30a64ae&is_raw=true&size=extra-large

in vagrant : 

```
Topic: bodhi.update.edit
Headers: {
    "fedora_messaging_rpm_gl2ps": true,
    "fedora_messaging_schema": "bodhi.update.edit.v1",
    "fedora_messaging_severity": 20,
    "fedora_messaging_user_cverna": true,
    "sent-at": "2020-05-25T07:34:02+00:00"
}
Body: {
    "agent": "cverna",
    "new_bugs": [],
    "update": {
        "alias": "FEDORA-2020-db7697713b",
        "autokarma": true,
        "autotime": true,
        "bugs": [],
        "builds": [
            {
                "id": 257971,
                "nvr": "gl2ps-1.4.1-6.fc33",
                "signed": false,
                "type": "rpm"
            }
        ],
        "close_bugs": true,
        "comments": [
            {
                "id": 1317535,
                "karma": 0,
                "text": "This update has been submitted for testing by cverna. ",
                "timestamp": "2020-05-20 10:28:07",
                "user": {
                    "id": 91,
                    "name": "bodhi"
                }
            },
            {
                "id": 1317536,
                "karma": 0,
                "text": "This update's test gating status has been changed to 'greenwave_failed'.",
                "timestamp": "2020-05-20 10:28:10",
                "user": {
                    "id": 91,
                    "name": "bodhi"
                }
            },
            {
                "id": 1317537,
                "karma": 0,
                "text": "This update's test gating status has been changed to 'waiting'.",
                "timestamp": "2020-05-20 10:28:10",
                "user": {
                    "id": 91,
                    "name": "bodhi"
                }
            },
            {
                "id": 1317538,
                "karma": 0,
                "text": "cverna edited this update.",
                "timestamp": "2020-05-20 10:32:12",
                "user": {
                    "id": 91,
                    "name": "bodhi"
                }
            },
            {
                "id": 1317539,
                "karma": 0,
                "text": "This update's test gating status has been changed to 'greenwave_failed'.",
                "timestamp": "2020-05-20 10:32:15",
                "user": {
                    "id": 91,
                    "name": "bodhi"
                }
            },
            {
                "id": 1317540,
                "karma": 0,
                "text": "cverna edited this update.",
                "timestamp": "2020-05-20 10:33:36",
                "user": {
                    "id": 91,
                    "name": "bodhi"
                }
            },
            {
                "id": 1317541,
                "karma": 0,
                "text": "cverna edited this update.",
                "timestamp": "2020-05-20 10:33:48",
                "user": {
                    "id": 91,
                    "name": "bodhi"
                }
            },
            {
                "id": 1317547,
                "karma": 0,
                "text": "cverna edited this update.",
                "timestamp": "2020-05-25 07:34:01",
                "user": {
                    "id": 91,
                    "name": "bodhi"
                }
            }
        ],
        "compose": null,
        "content_type": "rpm",
        "critpath": false,
        "date_approved": null,
        "date_modified": "2020-05-25 07:34:01",
        "date_pushed": null,
        "date_stable": null,
        "date_submitted": "2020-05-20 10:28:07",
        "date_testing": null,
        "display_name": "",
        "from_tag": null,
        "karma": 0,
        "locked": false,
        "meets_testing_requirements": false,
        "notes": "sada\n\n\n\n",
        "pushed": false,
        "release": {
            "id": 40,
            "long_name": "Fedora 33",
            "name": "F33"
        },
        "request": "testing",
        "require_bugs": true,
        "require_testcases": true,
        "requirements": null,
        "severity": "unspecified",
        "stable_days": 7,
        "stable_karma": 3,
        "status": "pending",
        "suggest": "unspecified",
        "test_cases": [],
        "test_gating_status": "greenwave_failed",
        "title": "gl2ps-1.4.1-6.fc33",
        "type": "unspecified",
        "unstable_karma": -3,
        "updateid": "FEDORA-2020-db7697713b",
        "url": "http://localhost:6543/updates/FEDORA-2020-db7697713b",
        "user": {
            "id": 3442,
            "name": "cverna"
        },
        "version_hash": "6801e410c07ada5b006fac30e5fd66c76e51fd68"
    }
}
```

I am not sure why the tests are still working (I have not looked at it yet), but if we want to go ahead with that change I think we need to create a v2 of the message schema