@mattiaverga 2020-05-25
I suppose the message tests are completely unreliable: the UpdateV1 schema description is completely different from the update object produced in the UpdateEditV1 message :hushed: (even before this PR)
Am I wrong?

Should I rewrite all object schemas from scratch and mark them V2?
What to do with schemas V1? Should they be removed, or we keep both V1 and V2?
Does that mean that every change we make to a model, also the message schema of that object should be changed?
Or we care only of fields described in the schema?